+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.4.16
 
-* support anonymization for Aperio format scanned by KFBIO
+* support anonymization for Aperio format scanned by KFBIO KF-PRO-400 and Motic Pro6
 
 ## 0.4.15